+++ src/BoardM5stackCore2.cpp
@@ -217,4 +217,41 @@ void BoardM5stackCore2::mainLoop()
   }
 }
 
+/**
+ * Set time
+ */
+void BoardM5stackCore2::setTime(String timestamp)
+{
+  RTC_TimeTypeDef RTCtime;
+  RTC_DateTypeDef RTCdate;
+
+  RTCdate.Year = timestamp.substring(0, 4).toInt();
+  RTCdate.Month = timestamp.substring(5, 7).toInt();
+  RTCdate.Date = timestamp.substring(8, 10).toInt();
+  RTCtime.Hours = timestamp.substring(11, 13).toInt();
+  RTCtime.Minutes = timestamp.substring(14, 16).toInt();
+  RTCtime.Seconds = timestamp.substring(17, 19).toInt();
+
+  M5.Rtc.SetTime(&RTCtime);
+  M5.Rtc.SetDate(&RTCdate);
+
+  BoardInterface::setTime(timestamp);
+}
+
+/**
+ * Sync NTP time
+ *
+ */
+void BoardM5stackCore2::ntpSync()
+{
+
+  syslog->println("Syncing NTP time.");
+
+  char ntpServer[] = "de.pool.ntp.org";
+  configTime(liveData->settings.timezone * 3600, liveData->settings.daylightSaving * 3600, ntpServer);
+  liveData->params.ntpTimeSet = true;
+
+  showTime();
+}
+
 #endif // BOARD_M5STACK_CORE2