+++ src/BoardM5stackCore2.h
@@ -20,4 +20,6 @@ class BoardM5stackCore2 : public Board320_240 {
     void Write1Byte(uint8_t Addr, uint8_t Data);
     uint8_t Read8bit(uint8_t Addr);
     static void eventDisplay(Event& e);
+    void setTime(String timestamp) override;
+    void ntpSync() override;
 };